Change the username of the user inside the container to 'marvin'
This essentially allows me to run the docker program inside the
container and solves the (docker related) relative mounts issue.

The gist of it here is that since the container is configured to use the
hosts' docker daemon, attempting to use relative mounts from within the
container does not work, since the user inside the container was called
`dev`, and not `marvin`.

Hopefully this fixes that!