Add the ability to tunnel DNS requests through a SOCKS5 proxy
The primary use case for this is to be able to tunnel a single Firefox
container's traffic through this proxy, including the DNS traffic.

In an environment where DNS requests are filtered or blocked (i.e.
Adguard, etc), it's useful to have a way to "turn it off" for specific
websites.